add scripts for accessing local emacs via ssh
I often ssh into another machine and do some work there. Emacs
(obviously) supports opening files on a remote machine (TRAMP), but it
is much more convenient to be able to open a file from the terminal
where one is working rather than doing C-x C-f in emacs and writing
the long ssh:///foo@bar:blabla path. On the local machine, that can be
done with emacsclient, provided we've set up emacs to accept incoming
connections (mostly just (server-start) in .emacs).

But we can achieve the same thing with a combination of a number of
nifty features in emacs and ssh: First, configure emacs to listen on a
tcp port (e.g. 9999, though it can actually be randomly chosen) rather
than a local unix socket. Next, forward a port on the remote machine
to port 9999 on the local machine. Using the ControlMaster feature of
openssh, that can be done via the same ssh connection as we're logged
in through, and conveniently, that also automatically records the
remote hosts we need to establish the port forwardings from (as well
as copy a little bit of extra info for use by the emacsclient on the
remote host).

So the prerequisites for making this work is something like this: In
.emacs, have something like

(setq server-use-tcp t
      server-port    9999)
(server-start)

In ~/.ssh/config, make sure to have

     ControlMaster auto
     ControlPath ~/.ssh/cm_sock/%r@%h:%p
     ControlPersist 5

stanzas for all the hosts you may log into and want to use emacsclient
on. We parse the socket names found in ~/.ssh/cm_sock, so if you
change that path and/or the %r@%h:%p format, you need to adapt the
ssh_emacs script.

Now, after ssh'ing to a remote host and having a running emacs
instance, just run the ssh_emacs script. That script loops over all
control sockets in ~/.ssh/cm_sock, and for each:

- asks the ssh control master to forward a (randomly chosen) port on
  the remote to the localhost:9999, or if such a forwarding already
  exists, reuse that

- makes a copy of the ~/.emacs.d/server/server file, modifying the
  first line to use the chosen remote port

- appends a third line to that file consisting of "$user@$host"

- copy this file to ~/.emacs.d/server/remote-server on the remote host

Now, the ec script is partly a convenient shorthand for emacsclient,
partly implements the last pieces to make all this work. First, it
checks whether there's a local emacs server running (by looking for a
~/.emacs.d/server/server file and sanity checking that), and if so, it
just acts as a normal emacsclient invocation. Otherwise, it looks for
a ~/.emacs.d/server/remote-server file. If that exists, it uses the
third line (the user@host appended by ssh_emacs) to munge the filename
arguments so that our local emacs can find them, and then it invokes
"emacsclient -f ~/.emacs.d/server/remote-server".

Finally, ecn is just a shorthand for "ec -n". It's a script rather
than just a shell alias to be usable as $EDITOR.

TODO: figure out how to cross-compile (just) emacsclient, and remove
bashisms from the ec script, so that I can use this on all the various
embedded boards I frequently log in to.