add askpass script
This is likely overengineering, but I did find myself tired of checking
in on the xterm, noting nothing gets written to the terminal, thinking
"hmm, something is going very slow or there's a bug somewhere causing
the build to hang", only to then remember "ahh, it might just be waiting
for me to enter my password". Also, in case something was indeed just
taking a long time, I don't like typing my password blindly.

So, if I just make sure to change the build scripts to use "sudo -A" and
pass this in SUDO_ASKPASS I should be good.